Use npx in makefile
Some of the makefile commands aren't working on some machines. I believe
that that's related to #524: when I revert that change, the commands
start working again on my machine. That said, it's not clear what the
specific issue is. We've discussed a couple of possibilities. I'm
running make v3, so maybe it's a matter of make v3 vs. v4? Or maybe of
GNU make vs. BSD make? Either way, using npx should work on all
machines.