Do not reuse database connection to drop database in migration tests
This is a related change to #583. We saw there that there might be an
issue if a database connection is used to migrate, then is later used to
drop the database. Instead, here we use a connection to complete a
single batch of migrations. If the connection drops the database, it
will do so before migrating.

This approach will also clean up the connection: migrator.destroy() will
be called (via withDatabase()).