Make Authorization header checks tighter (#849)
Include trailing space in checks for Basic and Bearer prefixes in
Authorization headers.  This space is already assumed to exist in
subsequent parsing code.

An extra check is added for requests which _provide_ the Authorization
header, but in an unrecognised format.  This ensures that a 401 header
is still sent in preference to 404, as per #329.