Option: block direct access to .value (#1287)
Guard against none-unsafe derefences of Option values.

This change requires implementation of support for Ramda's `.equals()` function for `Option.of(...)`, which incidentally worked previously due to the underlying implementation.  The main motivation for implementing `.equals()` support is that there are various places in the code which use `Ramda.isEmpty()` instead of `Option.isEmpty()`.  These are likely mistakes, but there is no obvious, elegant way to detect these mistakes.