task: remove remapping of node:util.promisify() (#1303)
Reading the current task code, it's not immediately clear why `task.promisify` is used in place of node's built-in `promisify` function. Digging deeper, it turns out that this is exactly the same function. This PR aims to save future developers from the overhead of understanding that `task.promisify === require('node:util').promisify`.