stderr.write(): add newline suffix consistently
Noted while investigating #1403

Some logging is not followed by a newline character, so it is concatenated with the next output to stderrr.

This commit adds newline characters to calls to process.stderr.write() where missing, and standardises their inclusion.

It's unclear why `process.stderr.write()` is used in preference to `console.error()`