Standardise responses for invalid new passwords (#1411)
The fix is pushed deeper into the `util/crypto.js` code than where validation previously occurred.  This means that validation only needs to be performed in a single place, rather than every point of use.  This in turn means it should be harder to forget validation & edge cases in future use of `hashPassword()`.

Closes #1407