resources/forms: remove outdated comment (#1529)
This comment was introduced in 4c9fbf9f0036b7320cfafeaad0a8aca4e24b23b6, when the proceeding lines read:

	formResource('/projects/:projectId/forms/:id', (Form, params, withXml = false, options = QueryOptions.none) =>
	  ((withXml === true) ? Form.getWithXmlByProjectAndXmlFormId : Form.getByProjectAndXmlFormId)(params.projectId, params.id, options)
	    .then(getOrNotFound));

The comment no longer seems relevant to the code that follows it, which is now much simpler than it was originally.

This commit also adds a newline, breaking apart two resource definitions, in keeping with the rest of the file.