tests: do not use getOrNotFound() (#1533)
Replace getOrNotFound uses with `(o) => o.get()`

In API code `getOrNotFound()` makes sense, as it generates a `Problem` which is converted to an HTTP 404 response.

In test code, the `Problem` is not required, and while the code may look neater, it obscures stacktraces and makes test failures harder to understand.

Closes https://github.com/getodk/central/issues/1162