Forms.publish(): remove already-published check (#1576)
The existing check seems to protect against incorrect function usage, but does not actually check anything.

I tried to write a failing test case and fixing the check, but it seemed very low-level and likely brittle, for little or no gain.

Re-introduce a working check with an associated test case if this proves to be an issue in the future.

Closes #1575