docs: make info.version more visible
In the `info` section, then `description` is about 500 lines long.  This makes it a little tricky to see other properties of the `info` section.

This PR moves the `description` to the end of the `info` section.

Noted while reviewing https://github.com/getodk/central-backend/pull/1633 and trying to understand what the `license` property belonged to.