test/knex-migrations: delete skipped tests
Mocha intends `.skip()` to be used to mark a test as "pending": https://mochajs.org/#pending-tests

These tests are not pending - they are broken and unlikely to be fixed, as they test migrations which have already been released and so should not be changed except in extreme circumstances.  In such circumstances, they could be retrieved from git history.

It's not useful to keep these tests as examples for future test writers, as in general, relying on application code to either write or test a migration is prone to breaking in exactly the way that these tests have.

There is a more recent pattern for writing tests for database migrations in `test/db-migrations/`.